extract queryExtension and toConfiguredExtensions from shared to web
